pyproject.toml
[tool.poetry]
name = "multi-translate"
version = "0.1.0"
description = "Translate api using multiple external apis"
authors = ["Reuben Thomas-Davis <reubenestd@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
sqlalchemy = "^1.3.17"
fastapi = "^0.54.2"
databases = {extras = ["postgresql"], version = "^0.3.2"}
graphene = "^2.1.8"
httpx = "^0.13.1"
uvicorn = "^0.11.5"
google-cloud-translate = "^2.0.1"
boto3 = "^1.13.19"
grpcio = "^1.29.0"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.4.2"
pytablewriter = "^0.54.0"
gql = "^2.0.0"
importanize = "^0.7.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
